Set buttons from the new enum in OI

+++ OnyxTronix2017/src/org/usfirst/frc/team2231/robot/Buttons.java
@@ -17,7 +17,7 @@ public class Buttons {
 	    Button(int value) {
 	        this.value = value;
 	    }
-	    public double value() { return value; }
+	    public int value() { return value; }
 	}
 	 
 }
